Create a new hermod driver; this is intended to be used from other packages, and rarely called directly. If you are trying to run tasks on a cluster you do not need to call this!

Create a new hermod driver; this is intended to be used from other packages, and rarely called directly. If you are trying to run tasks on a cluster you do not need to call this!

hermod_driver(configure, submit, status, result, cancel, provision)

Arguments

configure

Function used to set core configuration for the driver. This function will be called from the hermod root directory (so getwd() will report the correct path). It can take any arguments, do any calculation and then must return any R object that can be serialised. The resulting configuration will be passed in as config to other driver functions.

submit

Submit a task to a cluster. This is run after the task is created (either automatically or manually) and takes as arguments the task id, the configuration, the path to the root.

status

Fetch a task status. Takes a vector of ids and returns a vector of the same length of statuses.

result

Fetch a task result. If needed, copies the result file into the current hermod root. Assume that a result is available (i.e., we've already checked that the task status is terminal)

cancel

Cancel one or more tasks. Takes a vector of task ids, and requests that these tasks are cancelled, returning a logical vector the same length indicating if cancellation was successful.

provision

Provision a library. Works with conan, and must accept method, config, path_root followed by ... to pass through to conan2::conan_configure. It is expected this function will trigger running conan to provision a library.