Manage environments

Create, update, list, view and delete environments.

hermod_environment_create(
  name = "default",
  sources = NULL,
  packages = NULL,
  overwrite = TRUE,
  root = NULL
)

hermod_environment_list(root = NULL)

hermod_environment_delete(name = "default", root = NULL)

hermod_environment_show(name = "default", root = NULL)

hermod_environment_exists(name = "default", root = NULL)

Arguments

name

Name of the environment. The name default is special; this is the environment that will be used by default (hence the name!).

sources

Files to source before starting a task. These will be sourced into the global (or execution) environment of the task. The paths must be relative to the hermod root, not the working directory.

packages

Packages to be attached before starting a task. These will be loaded with library() before the sources are sourced. If you need to attach a package after a script for some reason, just call library yourself within one of your source files.

overwrite

On environment creation, replace an environment with the same name.

root

A hermod root, or path to it. If NULL we search up your directory tree.

Value

Nothing, all are called for their side effects.